requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
PyMySQL==1.1.1
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0


